Ajustando tipo do indexador de ativos

+++ src/main/java/cadastro/ativos/api/dtos/AtivoRequest.java
@@ -21,7 +21,7 @@ public record AtivoRequest(
         BigDecimal valorBase,
 
         @NotNull(message = "Indexador é obrigatório")
-        int indexador,
+        BigDecimal indexador,
 
         @NotNull(message = "Apta à Negociação é obrigatória")
         Boolean aptaNegociacao,
